proxying of IdamApi with known user cache
