+++ lib-consumer/src/main/java/uk/gov/hmcts/libconsumer/CFTLibConfig.java
@@ -17,7 +17,7 @@ public class CFTLibConfig implements CFTLibConfigurer {
   @SneakyThrows
   @Override
   public void configure(CFTLib lib) {
-    lib.createProfile("a@b.com");
+    lib.createProfile("a@b.com","DIVORCE", "NO_FAULT_DIVORCE", "Submitted");
     lib.createRoles(
         "caseworker-divorce-courtadmin_beta",
         "caseworker-divorce-superuser",
