+++ rse-cft-lib/src/main/java/uk/gov/hmcts/rse/ccd/lib/BeanIsolator.java
@@ -2,11 +2,9 @@ package uk.gov.hmcts.rse.ccd.lib;
 
 import java.util.Arrays;
 
-import java.util.Set;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.BeansException;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.config.BeanDefinitionHolder;
 import org.springframework.beans.factory.config.BeanFactoryPostProcessor;
 import org.springframework.beans.factory.config.ConfigurableListableBeanFactory;
