+++ rse-cft-lib/src/main/java/uk/gov/hmcts/rse/ccd/lib/FlywayMigrator.java
@@ -7,7 +7,6 @@ import javax.sql.DataSource;
 import lombok.SneakyThrows;
 import org.flywaydb.core.Flyway;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.boot.autoconfigure.flyway.FlywayMigrationStrategy;
 import org.springframework.context.annotation.Configuration;
