+++ rse-cft-lib/src/main/java/uk/gov/hmcts/rse/ccd/lib/ForProjects.java
@@ -2,7 +2,6 @@ package uk.gov.hmcts.rse.ccd.lib;
 
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
-import lombok.RequiredArgsConstructor;
 
 @Retention(RetentionPolicy.RUNTIME)
 @interface ForProjects {
