+++ rse-cft-lib/src/main/java/uk/gov/hmcts/rse/ccd/lib/PropertiesInjector.java
@@ -1,9 +1,7 @@
 package uk.gov.hmcts.rse.ccd.lib;
 
 import java.util.Properties;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.context.event.ApplicationEnvironmentPreparedEvent;
-import org.springframework.boot.web.server.LocalServerPort;
 import org.springframework.context.ApplicationListener;
 import org.springframework.core.env.ConfigurableEnvironment;
 import org.springframework.core.env.PropertiesPropertySource;
