+++ rse-cft-lib/src/main/java/uk/gov/hmcts/rse/ccd/lib/api/CFTLib.java
@@ -2,7 +2,9 @@ package uk.gov.hmcts.rse.ccd.lib.api;
 
 import com.auth0.jwt.JWT;
 import com.auth0.jwt.algorithms.Algorithm;
-import java.io.File;
+import io.jsonwebtoken.Jwts;
+import io.jsonwebtoken.SignatureAlgorithm;
+import io.jsonwebtoken.impl.TextCodec;
 import java.nio.charset.Charset;
 import java.sql.Connection;
 import java.time.LocalDateTime;
@@ -11,8 +13,6 @@ import java.util.Date;
 import java.util.List;
 import javax.sql.DataSource;
 import lombok.SneakyThrows;
-import org.apache.commons.fileupload.FileItem;
-import org.apache.commons.fileupload.disk.DiskFileItem;
 import org.apache.commons.io.IOUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
@@ -20,11 +20,9 @@ import org.springframework.boot.context.event.ApplicationReadyEvent;
 import org.springframework.context.event.EventListener;
 import org.springframework.core.io.ByteArrayResource;
 import org.springframework.core.io.DefaultResourceLoader;
-import org.springframework.core.io.FileSystemResource;
 import org.springframework.core.io.ResourceLoader;
 import org.springframework.http.HttpEntity;
 import org.springframework.http.HttpHeaders;
-import org.springframework.http.HttpMethod;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.oauth2.jwt.Jwt;
@@ -32,9 +30,6 @@ import org.springframework.stereotype.Component;
 import org.springframework.util.LinkedMultiValueMap;
 import org.springframework.util.MultiValueMap;
 import org.springframework.web.client.RestTemplate;
-import org.springframework.web.multipart.MultipartFile;
-import org.springframework.web.multipart.commons.CommonsMultipartFile;
-import uk.gov.hmcts.ccd.definition.store.excel.endpoint.ImportController;
 import uk.gov.hmcts.ccd.definition.store.repository.SecurityClassification;
 import uk.gov.hmcts.ccd.definition.store.repository.model.UserRole;
 import uk.gov.hmcts.ccd.definition.store.rest.endpoint.UserRoleController;
@@ -53,26 +48,24 @@ public class CFTLib {
   UserProfileEndpoint userProfile;
 
   @Autowired
-  ImportController importController;
-
-  @Autowired
   CFTLibConfigurer configurer;
 
   @Value("http://localhost:${server.port}")
   private String baseUrl;
 
+  @SneakyThrows
   @EventListener(ApplicationReadyEvent.class)
   public void configure() {
     configurer.configure(this);
   }
 
-  public void createProfile(String id) {
+  public void createProfile(String id, String jurisdiction, String caseType, String state) {
     var p = new UserProfile();
     p.setId(id);
     p.setId(id);
-    p.setWorkBasketDefaultJurisdiction("DIVORCE");
-    p.setWorkBasketDefaultCaseType("NO_FAULT_DIVORCE");
-    p.setWorkBasketDefaultState("Submitted");
+    p.setWorkBasketDefaultJurisdiction(jurisdiction);
+    p.setWorkBasketDefaultCaseType(caseType);
+    p.setWorkBasketDefaultState(state);
     userProfile.populateUserProfiles(List.of(p), "banderous");
   }
 
@@ -132,5 +125,14 @@ public class CFTLib {
         .withClaim("tokenName", "access_token")
         .withExpiresAt(Date.from(LocalDateTime.now().plusDays(100).toInstant(ZoneOffset.UTC)))
         .sign(Algorithm.HMAC256("a secret"));
+  }
+
+  public static String generateDummyS2SToken(String serviceName) {
+    return Jwts.builder()
+        .setSubject(serviceName)
+        .setIssuedAt(new Date())
+        .signWith(SignatureAlgorithm.HS256, TextCodec.BASE64.encode("AA"))
+        .compact();
+
   }
 }
