+++ rse-cft-lib/src/main/java/uk/gov/hmcts/rse/ccd/lib/api/CFTLibConfigurer.java
@@ -1,5 +1,5 @@
 package uk.gov.hmcts.rse.ccd.lib.api;
 
 public interface CFTLibConfigurer {
-  void configure(CFTLib lib);
+  void configure(CFTLib lib) throws Exception;
 }
